Update 2/28
Updated Editor scripts for the Spell Editor. It works almost exactly as I need it to now.
Integrated BetterEvents, which will allow more flexibility when creating effect functions (multi-params, yes!).
Fixed a minor issue where effects were being checked twice. This was not affecting gameplay in any way sense the conditions for effect proc would not change (and we have no random effects yet), but it could become a problem.